refactor: Move `ProviderInstaller` interface into separate file, refactor `Init` interface to embed it.
